customer_order/requirements.txt
google-auth==2.22.0
social-auth-app-django==5.2.0
PyJWT==2.8.0